Pathfinding robot size (#16)
This is so the robot does not crash into a wall whilst trying to reach
it's designated POI. Now the POI's coordinates moves back relative to
the robot's size and the POI's angle ensuring that the robot does not it
an obstacle. I also took the time to ensure the safety of the
tokenReader method by choosing the default option if there are no
options chosen.